// src/enums/commonEnums.ts
export enum Roles {
  ADMIN = "admin",
  USER = "user",
}

export enum ApprovalRequestStatus {
  PENDING = "pending",
  APPROVED = "approved",
  REJECTED = "rejected",
}

export enum ApprovalRequestType {
  SELLER_APPLICATION = "seller_application",
  ARTWORK_SUBMISSION = "artwork_submission",
}